feat(Layout): add the container layout component
