+++ packages/hedgehog-lab/src/pages/Main/Main.tsx
@@ -1,15 +1,13 @@
 import React, {useEffect, useState} from "react";
 import Qs from 'qs';
-import {Grid, styled} from '@mui/material';
+import {Grid} from '@mui/material';
 import {compiler} from "../../compiler";
 import Results from "../../components/Results/Results";
 import Footer from "../../components/DeprecatedCode/Footer/Footer";
 import YourCode from "../../components/YourCode/YourCode";
 import {queryCache} from "react-query";
-import {useRecoilValue, useSetRecoilState} from "recoil";
+import {useSetRecoilState} from "recoil";
 import {compilerReFetchState, editorCodeState} from "../../components/YourCode/RYourCodeStates";
-import {sideBarWidth} from "../../components/YourCode/Config/SideBar";
-import {sideBarOpenState} from "../../components/Layout/RLayoutStates";
 
 const DEFAULT_SOURCE = `//write your code here
 print("hello world")
@@ -17,29 +15,10 @@ print("hello world")
 
 const lastRunningCode = localStorage.getItem('lastRunningCode')
 
-const MainContent = styled('main', {shouldForwardProp: (prop) => prop !== 'open'})<{
-    open?: boolean;
-}>(({theme, open}) => ({
-    flexGrow: 1,
-    transition: theme.transitions.create('margin', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginLeft: `-${sideBarWidth}px`,
-    ...(open && {
-        transition: theme.transitions.create('margin', {
-            easing: theme.transitions.easing.easeOut,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-        marginLeft: 0,
-    }),
-}));
-
 
 const Main = (): React.ReactElement => {
     const setEditorCode = useSetRecoilState<string>(editorCodeState)
     const setCompilerReFetch = useSetRecoilState<boolean>(compilerReFetchState);
-    const sideBarOpen = useRecoilValue(sideBarOpenState)
     // If auto_run=true, then hedgehog lab will run the script automatically after loading the code
     const [autoRun, setAutoRun] = useState<boolean>(false)
 
@@ -74,7 +53,7 @@ const Main = (): React.ReactElement => {
     }, [autoRun])
 
     return (
-        <MainContent open={sideBarOpen}>
+        <>
             <Grid container>
                 <Grid item xs={12}>
                     <Grid
@@ -101,7 +80,7 @@ const Main = (): React.ReactElement => {
             </Grid>
 
             <Footer/>
-        </MainContent>
+        </>
     );
 };
