+++ packages/hedgehog-lab/src/pages/Settings/Account.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {Box, Button, Card, Container, Grid, Typography} from "@mui/material";
+import {Box, Button, Card, Grid, Typography} from "@mui/material";
 import BaseForm from "../../components/Form/Base/BaseForm/BaseForm";
 import {useRecoilState} from "recoil";
 import {authActionState} from "../../components/User/Auth/RAuthStates";
@@ -27,45 +27,43 @@ const Account = (): React.ReactElement => {
 
     return (
         <>
-            <Container maxWidth={'lg'} sx={{mt: '30px'}}>
-                <Box textAlign={"center"} mb={'30px'}>
-                    <Typography variant={'h2'}>
-                        Personal information
-                    </Typography>
+            <Box textAlign={"center"} mb={'30px'}>
+                <Typography variant={'h2'}>
+                    Personal information
+                </Typography>
 
-                    <Typography variant={'h6'}>
-                        Your personal information and preferences in various HHLAB services
-                    </Typography>
-                </Box>
+                <Typography variant={'h6'}>
+                    Your personal information and preferences in various HHLAB services
+                </Typography>
+            </Box>
 
-                <BaseForm>
-                    <Card variant={"outlined"}>
-                        <Box m={'20px'}>
-                            {accountForm.map((item, index) => {
-                                return (
-                                    <Grid key={index} container spacing={2} mt={index > 0 ? '20px' : '0'}>
-                                        <Grid item xs={2} alignSelf={'center'}>
-                                            <Typography variant={'body1'}>
-                                                {item.text}
-                                            </Typography>
-                                        </Grid>
-                                        <Grid item xs={10}>
-                                            {item.render}
-                                        </Grid>
+            <BaseForm>
+                <Card variant={"outlined"}>
+                    <Box m={'20px'}>
+                        {accountForm.map((item, index) => {
+                            return (
+                                <Grid key={index} container spacing={2} mt={index > 0 ? '20px' : '0'}>
+                                    <Grid item xs={2} alignSelf={'center'}>
+                                        <Typography variant={'body1'}>
+                                            {item.text}
+                                        </Typography>
                                     </Grid>
-                                )
-                            })}
+                                    <Grid item xs={10}>
+                                        {item.render}
+                                    </Grid>
+                                </Grid>
+                            )
+                        })}
 
-                            <Box textAlign={'end'}>
-                                <Button variant={"contained"} sx={{mt: '20px'}} type={"submit"}
-                                        onClick={() => setAuthAction('settingAccount')}>
-                                    Update
-                                </Button>
-                            </Box>
+                        <Box textAlign={'end'}>
+                            <Button variant={"contained"} sx={{mt: '20px'}} type={"submit"}
+                                    onClick={() => setAuthAction('settingAccount')}>
+                                Update
+                            </Button>
                         </Box>
-                    </Card>
-                </BaseForm>
-            </Container>
+                    </Box>
+                </Card>
+            </BaseForm>
         </>
     )
 }
